requirements.txt
requests==2.11.1
vcrpy==1.10.3
pytest==3.0.3